Moving auth functions to auth package
